Changing Group to numbers
Great job using JavaScript + HTML. We can chat a bit more about what you are building so I can give you more help. One detail I would change here is to use group as numbers instead of strings if you plan to only use numbers in the future. Numbers is a good if you need to sort/organize the group numbers later on. Also can be easier to validate which I think is what you want are doing in the next lines of code.